separate example replays list into separate file
